#6 feat: 완료페이지 비상금 페이지에서도 사용가능하도록 수정
